face-recognition-backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
face_recognition==1.3.0
opencv-python-headless==4.8.0.74
numpy==1.25.1
Pillow==10.0.0
